requirements/base.txt
flask==2.1.2
sqlalchemy==1.4.39
marshmallow==3.17.0
pyyaml==6.0
mysqlclient==2.1.1
pyjwt==2.4.0